Final Project 1 Submission for Java Dev class
We were instructed to make a mockup summary report for our company, and in it show Psuedocode, a flowchart, and an analysis of how we are implementing Object Oriented Programming practices in our work. 
The Pet.java file is the code, for the pet class only, not a final working application. Creates the pet class and methods for checking in a pet (cat or dog) to the pet boarding business.